opts_read_geojson {yyjsonr}	R Documentation

Options for reading in GeoJSON

Description

Options for reading in GeoJSON

Usage

opts_read_geojson(
  type = c("sf", "sfc"),
  property_promotion = c("string", "list"),
  property_promotion_lgl = c("integer", "string")
)

Arguments

type
'sf' or 'sfc'

property_promotion
What is the most general container type to use when properties differ across a FEATURECOLLECTION? E.g. if the property exists both as a numeric and a string, should all values be promoted to a 'string', or contained as different types in a 'list'. Default: 'string' will behave like geojsonsf package.

property_promotion_lgl
when property_promotion = "string" should logical values become words (i.e. "TRUE"/"FALSE") or integers (i.e. "1"/"0"). Default: "integer" in order to match geojsonsf package

Value

Named list of options specific to reading GeoJSON

Examples

# Create a set of options to use when reading geojson
opts_read_geojson()

[Package yyjsonr version 0.1.21 Index]