read_json_conn {yyjsonr}	R Documentation

Parse JSON from an R connection object.

Description

Currently, this is not very efficient as the entire contents of the connection are read into R as a string and then the JSON parsed from there.

Usage

read_json_conn(conn, opts = list(), ...)

Arguments

conn
connection object. e.g. url('https://jsonplaceholder.typicode.com/todos/1')

opts
Named list of options for parsing. Usually created by opts_read_json()

...
Other named options can be used to override any options in opts. The valid named options are identical to arguments to opts_read_json()

Details

For plain text files it is faster to use read_json_file().

Value

R object

See Also

Other JSON Parsers: read_json_file(), read_json_raw(), read_json_str(), read_ndjson_file(), read_ndjson_raw(), read_ndjson_str()

Examples

if (interactive()) {
  read_json_conn(url("https://api.github.com/users/hadley/repos"))
}

[Package yyjsonr version 0.1.21 Index]