read_json_str {yyjsonr}	R Documentation

Convert JSON in a character string to R

Description

Convert JSON in a character string to R

Usage

read_json_str(str, opts = list(), ...)

Arguments

str
a single character string

opts
Named list of options for parsing. Usually created by opts_read_json()

...
Other named options can be used to override any options in opts. The valid named options are identical to arguments to opts_read_json()

Value

R object

See Also

Other JSON Parsers: read_json_conn(), read_json_file(), read_json_raw(), read_ndjson_file(), read_ndjson_raw(), read_ndjson_str()

Examples

read_json_str("4294967297", opts = opts_read_json(int64 = 'string'))

[Package yyjsonr version 0.1.21 Index]