write_json_raw {yyjsonr}	R Documentation

Convert R object to a raw vector of JSON data

Description

Convert R object to a raw vector of JSON data

Usage

write_json_raw(x, opts = list(), ...)

Arguments

x
the object to be encoded

opts
Named list of serialization options. Usually created by opts_write_json()

...
Other named options can be used to override any options in opts. The valid named options are identical to arguments to opts_write_json()

Value

Raw vector containing the JSON

See Also

Other JSON Serializer: write_json_file(), write_json_str(), write_ndjson_file(), write_ndjson_raw(), write_ndjson_str()

Examples

js <- write_json_raw(head(iris, 3))
js
read_json_raw(js)

[Package yyjsonr version 0.1.21 Index]