write_ndjson_str {yyjsonr}	R Documentation

Write list or data.frame object to NDJSON in a string

Description

For list input, each element of the list is written as a single JSON string. For data.frame input, each row of the data.frame is written as aJSON string.

Usage

write_ndjson_str(x, opts = list(), ...)

Arguments

x
data.frame or list to be written as multiple JSON strings

opts
Named list of serialization options. Usually created by opts_write_json()

...
Other named options can be used to override any options in opts. The valid named options are identical to arguments to opts_write_json()

Value

String containing multiple JSON strings separated by newlines.

See Also

Other JSON Serializer: write_json_file(), write_json_raw(), write_json_str(), write_ndjson_file(), write_ndjson_raw()

Examples

write_ndjson_str(head(mtcars))

[Package yyjsonr version 0.1.21 Index]